requirements.txt
# Server-side dependencies only — no ML/GPU packages (those live in worker/requirements.txt)

# Web framework
fastapi
uvicorn[standard]
websockets
python-multipart

# Database
sqlalchemy

# Auth
authlib
httpx
python-jose[cryptography]
passlib

# Utilities
pyyaml
anthropic
